BL-6132 Src folder structure refactor and package build fix (#3)
* Src folder structure refactor

* Production package should not contain local smmt service module. It is used for debug purpose only
